+++ src/com/dhbw/MorseProject/send/events/IEncoderErrorListener.java
@@ -1,4 +1,4 @@
-package com.dhbw.MorseProject.send;
+package com.dhbw.MorseProject.send.events;
 
 public interface IEncoderErrorListener {
     void run(String message);
