+++ src/com/dhbw/MorseProject/send/events/IEncoderFinishedListener.java
@@ -1,4 +1,4 @@
-package com.dhbw.MorseProject.send;
+package com.dhbw.MorseProject.send.events;
 
 public interface IEncoderFinishedListener {
     void run();
